% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shap_summary.R
\name{plot_shap_summary}
\alias{plot_shap_summary}
\title{Plot SHAP Summary}
\usage{
plot_shap_summary(shap_matrix, data)
}
\arguments{
\item{shap_matrix}{The matrix output from survey_shap().}

\item{data}{The original dataframe used for training (to provide feature values).}
}
\value{
A ggplot2 object.
}
\description{
Plot SHAP Summary
}
