% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_shap.R
\name{survey_shap}
\alias{survey_shap}
\title{Extract SHAP Values from a Survey-Weighted XGBoost Model}
\usage{
survey_shap(sxgb_model)
}
\arguments{
\item{sxgb_model}{The list output from the \code{survey_xgboost()} function.}
}
\value{
A matrix of SHAP values detailing the marginal contribution of each
feature to the final prediction for every observation.
}
\description{
This function cracks open a trained survey-weighted XGBoost model and calculates
the SHapley Additive exPlanations (SHAP values) for every respondent.
}
